# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Klinik Sehat API",
        "description": "REST API untuk aplikasi manajemen klinik. Mencakup manajemen dokter, jadwal, reservasi pasien, dan rekam medis.",
        "contact": {
            "name": "Klinik Sehat Dev Team",
            "email": "admin@klinik.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api",
            "description": "Local Development Server"
        }
    ],
    "paths": {
...
        "/reservasis": {
            "get": {
                "tags": [
                    "Reservasi"
                ],
                "summary": "Daftar reservasi milik sendiri",
                "description": "Pasien hanya dapat melihat reservasi milik sendiri. Support filter status dan pagination.",
                "operationId": "b804709c05b64f9770b6f410263b4002",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Filter status",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "pending",
                                "dikonfirmasi",
                                "selesai",
                                "dibatalkan"
                            ]
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Daftar reservasi"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Reservasi"
                ],
                "summary": "Buat reservasi baru",
                "description": "user_id diambil otomatis dari JWT token. nomor_antrian digenerate otomatis.",
                "operationId": "fa7d4731e50377f299d6ff5b9fba94ce",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "dokter_id",
                                    "jadwal_id",
                                    "tanggal_reservasi",
                                    "keluhan"
                                ],
                                "properties": {
                                    "dokter_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "jadwal_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "tanggal_reservasi": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2026-06-10"
                                    },
                                    "keluhan": {
                                        "type": "string",
                                        "example": "Demam tinggi selama 3 hari",
                                        "minLength": 10
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Reservasi berhasil dibuat",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Reservasi berhasil dibuat."
                                        },
                                        "data": {
                                            "properties": {
                                                "nomor_antrian": {
                                                    "type": "string",
                                                    "example": "ANT-K7PQ2M"
                                                },
                                                "status": {
                                                    "type": "string",
                                                    "example": "pending"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "422": {
                        "description": "Validasi gagal"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/reservasis/{id}": {
            "get": {
                "tags": [
                    "Reservasi"
                ],
                "summary": "Detail reservasi milik sendiri",
                "operationId": "0aabdc6a9c68ca4204c80ec84a92da22",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detail reservasi beserta rekam medis jika sudah ada"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Reservasi tidak ditemukan"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Reservasi"
                ],
                "summary": "Batalkan reservasi (pasien only)",
                "description": "Pasien hanya bisa membatalkan reservasi yang masih berstatus 'pending'.",
                "operationId": "cb9a05645b50c60706bda3c6c4f71653",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "status": {
                                        "type": "string",
                                        "example": "dibatalkan",
                                        "enum": [
                                            "dibatalkan"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Reservasi berhasil dibatalkan"
                    },
                    "422": {
                        "description": "Status bukan pending, tidak bisa dibatalkan"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Reservasi"
                ],
                "summary": "Hapus reservasi (hanya yang masih pending)",
                "operationId": "4ca2c7b43f23e652d2d25c549099bc3f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Reservasi berhasil dihapus"
                    },
                    "404": {
                        "description": "Tidak ditemukan atau bukan milik sendiri"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/admin/reservasis": {
            "get": {
                "tags": [
                    "Admin"
                ],
                "summary": "Semua reservasi seluruh pasien (admin only)",
                "operationId": "53429823529e90001c86ec949c7dee92",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "pending",
                                "dikonfirmasi",
                                "selesai",
                                "dibatalkan"
                            ]
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Cari nama pasien",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "dokter_id",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 15
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Semua reservasi"
                    },
                    "401": {
                        "description": "Unauthenticated / API Key salah"
                    },
                    "403": {
                        "description": "Bukan admin"
                    }
                },
                "security": [
                    {
                        "bearerAuth": [],
                        "apiKeyAuth": []
                    }
                ]
            }
        },
        "/admin/reservasis/{id}/status": {
            "patch": {
                "tags": [
                    "Admin"
                ],
                "summary": "Update status reservasi (admin only)",
                "operationId": "2b25789f653439b303da0bfed69c4ca7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "status": {
                                        "type": "string",
                                        "example": "dikonfirmasi",
                                        "enum": [
                                            "pending",
                                            "dikonfirmasi",
                                            "selesai",
                                            "dibatalkan"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Status berhasil diperbarui"
                    },
                    "403": {
                        "description": "Bukan admin"
                    },
                    "404": {
                        "description": "Reservasi tidak ditemukan"
                    }
                },
                "security": [
                    {
                        "bearerAuth": [],
                        "apiKeyAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Masukkan JWT token dari response login. Format: Bearer {token}",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            },
            "apiKeyAuth": {
                "type": "apiKey",
                "description": "API Key dari response login/register. Wajib untuk endpoint admin.",
                "name": "X-API-KEY",
                "in": "header"
            }
        }
    },
...
}